+++ ai-engine/src/agents/asset_converter.py
@@ -8,7 +8,7 @@ import json
 import base64
 from pathlib import Path
 from langchain.tools import tool
-from models.smart_assumptions import (
+from src.models.smart_assumptions import (
     SmartAssumptionEngine, FeatureContext, ConversionPlanComponent
 )
