+++ ai-engine/src/agents/logic_translator.py
@@ -8,7 +8,7 @@ import logging
 import json
 import re
 from langchain.tools import tool
-from models.smart_assumptions import (
+from src.models.smart_assumptions import (
     SmartAssumptionEngine, FeatureContext, ConversionPlanComponent
 )
