+++ ai-engine/src/agents/qa_validator.py
@@ -9,7 +9,7 @@ import json
 from datetime import datetime
 from pathlib import Path
 from langchain.tools import tool
-from models.smart_assumptions import (
+from src.models.smart_assumptions import (
     SmartAssumptionEngine, FeatureContext, ConversionPlanComponent, AssumptionReport
 )
